+++ app/scenes/Trash/index.tsx
@@ -4,12 +4,11 @@ import * as React from "react";
 import { useTranslation } from "react-i18next";
 import Button from "~/components/Button";
 import Empty from "~/components/Empty";
-import Flex from "~/components/Flex";
 import Heading from "~/components/Heading";
 import PaginatedDocumentList from "~/components/PaginatedDocumentList";
 import Scene from "~/components/Scene";
 import Subheading from "~/components/Subheading";
-import { permanentlyDeleteDocumentsInTrash } from "~/actions/definitions/trash";
+import { permanentlyDeleteDocumentsInTrash } from "~/actions/definitions/documents";
 import useActionContext from "~/hooks/useActionContext";
 import useStores from "~/hooks/useStores";
 
@@ -18,10 +17,11 @@ function Trash() {
   const { documents } = useStores();
   const context = useActionContext();
   return (
-    <Scene icon={<TrashIcon />} title={t("Trash")}>
-      <Flex align="baseline" justify="space-between">
-        <Heading>{t("Trash")}</Heading>
-        {documents.deleted.length > 0 && (
+    <Scene
+      icon={<TrashIcon />}
+      title={t("Trash")}
+      actions={
+        documents.deleted.length > 0 && (
           <Button
             neutral
             action={permanentlyDeleteDocumentsInTrash}
@@ -29,8 +29,10 @@ function Trash() {
           >
             Empty
           </Button>
-        )}
-      </Flex>
+        )
+      }
+    >
+      <Heading>{t("Trash")}</Heading>
       <PaginatedDocumentList
         documents={documents.deleted}
         fetch={documents.fetchDeleted}
