+++ shared/i18n/locales/en_US/translation.json
@@ -66,6 +66,8 @@
   "Delete {{ documentName }}": "Delete {{ documentName }}",
   "Permanently delete": "Permanently delete",
   "Permanently delete {{ documentName }}": "Permanently delete {{ documentName }}",
+  "Empty": "Empty",
+  "Permanently delete documents in trash": "Permanently delete documents in trash",
   "Comments": "Comments",
   "History": "History",
   "Insights": "Insights",
@@ -101,8 +103,6 @@
   "Select a workspace": "Select a workspace",
   "New workspace": "New workspace",
   "Create a workspace": "Create a workspace",
-  "Empty": "Empty",
-  "Permanently delete all documents": "Permanently delete all documents",
   "Invite people": "Invite people",
   "Invite to workspace": "Invite to workspace",
   "Delete user": "Delete user",
@@ -945,7 +945,7 @@
   "Workspace name": "Workspace name",
   "You are creating a new workspace using your current account — <em>{{email}}</em>": "You are creating a new workspace using your current account — <em>{{email}}</em>",
   "To create a workspace under another email please sign up from the homepage": "To create a workspace under another email please sign up from the homepage",
-  "Deleted all the documents in Trash": "Deleted all the documents in Trash",
+  "Trash emptied": "Trash emptied",
   "Are you sure you want to permanently delete all the documents in Trash? This action is immediate and cannot be undone.": "Are you sure you want to permanently delete all the documents in Trash? This action is immediate and cannot be undone.",
   "Trash is empty at the moment.": "Trash is empty at the moment.",
   "A confirmation code has been sent to your email address, please enter the code below to permanently destroy your account.": "A confirmation code has been sent to your email address, please enter the code below to permanently destroy your account.",
